FPM-803
- Added the error message for when the returned object from the chooseImage method is null or undefined
